feat: reorganize inseri global object
